+++ local/testing/testdata/transform-and-upload.py
@@ -60,6 +60,8 @@ def create_edc_registry_asset_payload(registry_url_, asset_prop_id_):
         "@context": edc_context(),
         "@id": f"{asset_prop_id_}",  # DTR-EDC-instance-unique-ID
         "edc:properties": {
+            "dct:type": {"@id": "https://w3id.org/catenax/taxonomy#DigitalTwinRegistry"},
+            "cx-common:version": "3.0",
             "edc:description": "Digital Twin Registry Endpoint of IRS DEV",
             "edc:id": f"{asset_prop_id_}",  # DTR-EDC-instance-unique-ID
             "edc:type": "data.core.digitalTwinRegistry"
@@ -84,7 +86,8 @@ def edc_context():
         "edc": "https://w3id.org/edc/v0.0.1/ns/",
         "dcat": "https://www.w3.org/ns/dcat/",
         "odrl": "http://www.w3.org/ns/odrl/2/",
-        "dspace": "https://w3id.org/dspace/v0.8/"
+        "dspace": "https://w3id.org/dspace/v0.8/",
+        "cx-common": "https://w3id.org/catenax/ontology/common#"
     }