+++ CHANGELOG.md
@@ -8,6 +8,13 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 
 ## [Unreleased]
 
+### Fixed
+
+- IRS now searches for Digital Twin Registry contract offers by
+  type `dct:type`: `https://w3id.org/catenax/taxonomy#DigitalTwinRegistry`
+  or `edc:type`: `data.core.digitalTwinRegistry`. #616
+
+
 ## [5.1.2] - 2024-05-13
 
 ### Fixed